+++ Scripts/Ship/Ship.cs
@@ -10,7 +10,7 @@ public partial class Ship : Node2D
     [Export]
     public bool isMine = false;
 
-    [Export] public int health;
+    [Export] public float health;
     public float speed = 100f;
     public Area2D area2D;   
     public CollisionShape2D collisionShape;
